+++ packages/gestalt/src/Masonry.js
@@ -51,7 +51,6 @@ type Props<T> = {|
           from: number,
         }
       ) => void | boolean | {}),
-  onFinishedRendering?: () => void,
   scrollContainer?: () => HTMLElement,
   virtualBoundsTop?: number,
   virtualBoundsBottom?: number,
@@ -228,7 +227,7 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
   }
 
   componentDidUpdate(prevProps: Props<T>, prevState: State<T>) {
-    const { items, measurementStore, onFinishedRendering } = this.props;
+    const { items, measurementStore } = this.props;
 
     this.measureContainerAsync();
 
@@ -249,8 +248,6 @@ export default class Masonry<T> extends React.Component<Props<T>, State<T>> {
           hasPendingMeasurements,
         });
       });
-    } else if (onFinishedRendering) {
-      onFinishedRendering();
     }
   }