+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Minor
 
 - Box: Add ref forwarding (#431)
+- Masonry: Removed onFinishedRendering prop because better test alternatives could be used (#435)
 
 ### Patch